Moved some data + plotting
